Extract out feed label input component from `FeedListItem`
