+++ shared/src/commonMain/kotlin/dev/sasikanth/rss/reader/feeds/ui/FeedListItem.kt
@@ -18,20 +18,15 @@ package dev.sasikanth.rss.reader.feeds.ui
 import androidx.compose.animation.AnimatedContent
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.interaction.MutableInteractionSource
-import androidx.compose.foundation.interaction.collectIsFocusedAsState
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.requiredHeight
 import androidx.compose.foundation.layout.requiredSize
 import androidx.compose.foundation.layout.requiredWidth
 import androidx.compose.foundation.layout.sizeIn
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.text.KeyboardActions
-import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.rounded.MoreVert
 import androidx.compose.material3.Badge
@@ -39,13 +34,8 @@ import androidx.compose.material3.BadgedBox
 import androidx.compose.material3.Divider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextField
-import androidx.compose.material3.TextFieldDefaults
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.derivedStateOf
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -57,12 +47,11 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.platform.LocalFocusManager
-import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.unit.DpOffset
 import androidx.compose.ui.unit.dp
 import dev.sasikanth.rss.reader.components.DropdownMenu
 import dev.sasikanth.rss.reader.components.DropdownMenuItem
+import dev.sasikanth.rss.reader.components.FeedLabelInput
 import dev.sasikanth.rss.reader.components.image.AsyncImage
 import dev.sasikanth.rss.reader.core.model.local.Feed
 import dev.sasikanth.rss.reader.feeds.ui.FeedsSheetMode.Edit
@@ -77,7 +66,6 @@ import dev.sasikanth.rss.reader.resources.icons.Website
 import dev.sasikanth.rss.reader.resources.strings.LocalStrings
 import dev.sasikanth.rss.reader.share.LocalShareHandler
 import dev.sasikanth.rss.reader.ui.AppTheme
-import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 
 @Composable
@@ -290,84 +278,4 @@ private fun PinFeedIconButton(
       tint = pinnedIconColor
     )
   }
-}
-
-@Composable
-private fun FeedLabelInput(
-  value: String,
-  onFeedNameChanged: (String) -> Unit,
-  modifier: Modifier = Modifier,
-  enabled: Boolean = true
-) {
-  // Maintaining local state so that it updates the text field in the UI
-  // instantly and doesn't have any weird UI state issues.
-  //
-  // I probably can extract this out into the presenter, we would have to
-  // maintain a list of text field states that are derived from the feeds list
-  // but this seems like a good alternative.
-  //
-  var input by remember(value) { mutableStateOf(value) }
-  var inputModified by remember(value) { mutableStateOf(false) }
-
-  val focusManager = LocalFocusManager.current
-  val isInputBlank by derivedStateOf { input.isBlank() }
-  val interactionSource = remember { MutableInteractionSource() }
-  val isFocused by interactionSource.collectIsFocusedAsState()
-
-  fun onFeedNameChanged(clearFocus: Boolean = true) {
-    inputModified = input != value
-
-    if (!isInputBlank && inputModified) {
-      onFeedNameChanged.invoke(input)
-    }
-
-    if (clearFocus) {
-      focusManager.clearFocus()
-    }
-  }
-
-  LaunchedEffect(isFocused) {
-    if (!isFocused && !inputModified) {
-      input = value
-    }
-  }
-
-  LaunchedEffect(input) {
-    // Same as setting a debounce
-    delay(500)
-    onFeedNameChanged(clearFocus = false)
-  }
-
-  TextField(
-    modifier = modifier.requiredHeight(56.dp).fillMaxWidth(),
-    value = input,
-    onValueChange = { input = it },
-    keyboardOptions = KeyboardOptions(imeAction = ImeAction.Done, autoCorrect = false),
-    keyboardActions = KeyboardActions(onDone = { onFeedNameChanged() }),
-    singleLine = true,
-    textStyle = MaterialTheme.typography.titleMedium,
-    shape = RoundedCornerShape(16.dp),
-    enabled = enabled,
-    interactionSource = interactionSource,
-    colors =
-      TextFieldDefaults.colors(
-        focusedContainerColor = AppTheme.colorScheme.tintedSurface,
-        unfocusedContainerColor = AppTheme.colorScheme.tintedSurface,
-        disabledContainerColor = AppTheme.colorScheme.tintedBackground,
-        focusedIndicatorColor = Color.Transparent,
-        unfocusedIndicatorColor = Color.Transparent,
-        disabledIndicatorColor = Color.Transparent,
-        errorIndicatorColor = Color.Transparent,
-        disabledTextColor = AppTheme.colorScheme.textEmphasisHigh,
-        focusedTextColor = AppTheme.colorScheme.textEmphasisHigh,
-        unfocusedTextColor = AppTheme.colorScheme.textEmphasisHigh,
-      ),
-    placeholder = {
-      Text(
-        text = LocalStrings.current.feedNameHint,
-        style = MaterialTheme.typography.labelLarge,
-        color = AppTheme.colorScheme.tintedForeground.copy(alpha = 0.4f)
-      )
-    }
-  )
 }
